Change maps columns center_latitude and center_longitude to data type DECIMAL.

+++ db/schema/01_schema.sql
@@ -16,8 +16,8 @@ CREATE TABLE users (
 CREATE TABLE maps (
   id SERIAL PRIMARY KEY NOT NULL,
   zoom INTEGER,
-  center_latitude INTEGER,
-  center_longitude INTEGER
+  center_latitude DECIMAL,
+  center_longitude DECIMAL
 );
 
 -- Recreate Pins Table
